pyproject.toml
[tool.poetry]
name = "fetch-agents"
version = "0.1.0"
description = ""
authors = ["Fatumayattani <fyattani@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.12"
uagents = "^0.21.0"
openai = "^1.66.5"
python-dotenv = "^1.0.1"
requests = "^2.32.3"
tortoise-orm = "^0.24.2"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
